+++ compiler/Options.hs
@@ -35,7 +35,7 @@ data Options
         , enum_header :: Bool
         , allocator :: Maybe String
         , apply :: [ApplyOptions]
-        , apply_attribute :: Maybe String
+        , export_attribute :: Maybe String
         , jobs :: Maybe Int
         , no_banner :: Bool
         }
@@ -71,7 +71,7 @@ cpp = Cpp
     , enum_header = def &= name "e" &= help "Generate enums into a separate header file"
     , allocator = def &= typ "ALLOCATOR" &= help "Generate types using the specified  allocator"
     , apply = def &= typ "PROTOCOL" &= help "Generate Apply function overloads for the specified protocol only; supported protocols: compact, fast and simple"
-    , apply_attribute = def &= typ "ATTRIBUTE" &= help "Prefix the declarations of Apply functions with the specified C++ attribute/declspec"
+    , export_attribute = def &= typ "ATTRIBUTE" &= explicit &= name "apply-attribute" &= name "export-attribute" &= help "Prefix declarations for library export with the specified C++ attribute/declspec. apply-attribute is a deprecated synonym."
     , jobs = def &= opt "0" &= typ "NUM" &= name "j" &= help "Run NUM jobs simultaneously (or '$ncpus' if no NUM is not given)"
     , no_banner = def &= help "Omit the banner at the top of generated files"
     } &=
@@ -93,7 +93,6 @@ schema = Schema
     } &=
     name "schema" &=
     help "Output the JSON representation of the schema"
-
 
 mode :: Mode (CmdArgs Options)
 mode = cmdArgsMode $ modes [cpp, cs, schema] &=